rust: move staticlib to smart_keymap workspace member
