+++ src/lib.rs
@@ -92,86 +92,3 @@ pub mod init {
 
 #[cfg(custom_keymap)]
 include!(concat!(env!("OUT_DIR"), "/keymap.rs"));
-
-static mut KEYMAP: keymap::Keymap<init::KeyDefinitionsType> =
-    keymap::Keymap::new(init::KEY_DEFINITIONS, init::CONTEXT);
-
-/// Initialize the global keymap instance.
-#[allow(static_mut_refs)]
-#[no_mangle]
-pub extern "C" fn keymap_init() {
-    unsafe {
-        KEYMAP.init();
-    }
-}
-
-/// Register a keypress event to the global keymap instance.
-#[allow(static_mut_refs)]
-#[no_mangle]
-pub extern "C" fn keymap_register_input_keypress(keymap_index: u16) {
-    unsafe {
-        KEYMAP.handle_input(input::Event::Press { keymap_index });
-    }
-}
-
-/// Register a keyrelease event to the global keymap instance.
-#[allow(static_mut_refs)]
-#[no_mangle]
-pub extern "C" fn keymap_register_input_keyrelease(keymap_index: u16) {
-    unsafe {
-        KEYMAP.handle_input(input::Event::Release { keymap_index });
-    }
-}
-
-/// Run Keymap processing.
-///
-/// Should be called every ms.
-///
-/// The HID keyboard report is copied to the given buffer.
-///
-/// The `buf` report is for the HID boot keyboard.
-///
-/// # Safety
-///
-/// `buf` must be a valid pointer to a buffer of at least 8 bytes.
-#[allow(static_mut_refs)]
-#[no_mangle]
-pub unsafe extern "C" fn keymap_tick(buf: *mut u8) {
-    if buf.is_null() {
-        return;
-    }
-
-    unsafe {
-        KEYMAP.tick();
-
-        let report = KEYMAP.report_output().as_hid_boot_keyboard_report();
-        core::ptr::copy_nonoverlapping(report.as_ptr(), buf, report.len());
-    }
-}
-
-/// Copy the HID keyboard report to the given buffer.
-///
-/// It's better to use [keymap_tick] copy the report to the buffer,
-///  because the report won't change between [keymap_tick] calls.
-///
-/// # Safety
-///
-/// `buf` must be a valid pointer to a buffer of at least 8 bytes.
-#[allow(static_mut_refs)]
-#[no_mangle]
-pub unsafe extern "C" fn copy_hid_boot_keyboard_report(buf: *mut u8) {
-    if buf.is_null() {
-        return;
-    }
-
-    unsafe {
-        let report = keymap::KeymapOutput::new(KEYMAP.pressed_keys()).as_hid_boot_keyboard_report();
-        core::ptr::copy_nonoverlapping(report.as_ptr(), buf, report.len());
-    }
-}
-
-#[cfg(all(not(feature = "std"), feature = "staticlib"))]
-#[panic_handler]
-fn panic(_info: &core::panic::PanicInfo) -> ! {
-    loop {}
-}
